+++ include/bmfs/entry.h
