+++ src/bmfs-fuse.c
@@ -1,5 +1,5 @@
-#include "bmfs.h"
-#include "stdlib.h"
+#include <bmfs/bmfs.h>
+#include <bmfs/stdlib.h>
 
 #define FUSE_USE_VERSION 30
