+++ src/sspec-test.c
@@ -1,6 +1,5 @@
-#include "sspec.h"
-
 #include <assert.h>
+#include <bmfs/sspec.h>
 #include <errno.h>
 #include <stdlib.h>
 #include <string.h>
